+++ src/extension.ts
@@ -3,8 +3,9 @@ import axios, { isAxiosError } from "axios";
 import { getErrorMessage } from "coder/site/src/api/errors";
 import * as module from "module";
 import * as vscode from "vscode";
-import { makeCoderSdk, needToken } from "./api";
-import { errToStr } from "./api-helper";
+import { errToStr } from "./api/api-helper";
+import { CoderApi } from "./api/coderApi";
+import { needToken } from "./api/utils";
 import { Commands } from "./commands";
 import { CertificateError, getErrorDetail } from "./error";
 import { Remote } from "./remote";
@@ -64,21 +65,22 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 	// the plugin to poll workspaces for the current login, as well as being used
 	// in commands that operate on the current login.
 	const url = storage.getUrl();
-	const restClient = makeCoderSdk(
+	const client = CoderApi.create(
 		url || "",
 		await storage.getSessionToken(),
-		storage,
+		storage.output,
+		() => vscode.workspace.getConfiguration(),
 	);
 
 	const myWorkspacesProvider = new WorkspaceProvider(
 		WorkspaceQuery.Mine,
-		restClient,
+		client,
 		storage,
 		5,
 	);
 	const allWorkspacesProvider = new WorkspaceProvider(
 		WorkspaceQuery.All,
-		restClient,
+		client,
 		storage,
 	);
 
@@ -130,7 +132,7 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 					storage.getUrl(),
 				);
 				if (url) {
-					restClient.setHost(url);
+					client.setHost(url);
 					await storage.setUrl(url);
 				} else {
 					throw new Error(
@@ -144,11 +146,11 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 				// command currently always requires a token file.  However, if there is
 				// a query parameter for non-token auth go ahead and use it anyway; all
 				// that really matters is the file is created.
-				const token = needToken()
+				const token = needToken(vscode.workspace.getConfiguration())
 					? params.get("token")
 					: (params.get("token") ?? "");
 				if (token) {
-					restClient.setSessionToken(token);
+					client.setSessionToken(token);
 					await storage.setSessionToken(token);
 				}
 
@@ -212,7 +214,7 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 					storage.getUrl(),
 				);
 				if (url) {
-					restClient.setHost(url);
+					client.setHost(url);
 					await storage.setUrl(url);
 				} else {
 					throw new Error(
@@ -226,7 +228,7 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 				// command currently always requires a token file.  However, if there is
 				// a query parameter for non-token auth go ahead and use it anyway; all
 				// that really matters is the file is created.
-				const token = needToken()
+				const token = needToken(vscode.workspace.getConfiguration())
 					? params.get("token")
 					: (params.get("token") ?? "");
 
@@ -251,7 +253,7 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 
 	// Register globally available commands.  Many of these have visibility
 	// controlled by contexts, see `when` in the package.json.
-	const commands = new Commands(vscodeProposed, restClient, storage);
+	const commands = new Commands(vscodeProposed, client, storage);
 	vscode.commands.registerCommand("coder.login", commands.login.bind(commands));
 	vscode.commands.registerCommand(
 		"coder.logout",
@@ -319,8 +321,8 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 			if (details) {
 				// Authenticate the plugin client which is used in the sidebar to display
 				// workspaces belonging to this deployment.
-				restClient.setHost(details.url);
-				restClient.setSessionToken(details.token);
+				client.setHost(details.url);
+				client.setSessionToken(details.token);
 			}
 		} catch (ex) {
 			if (ex instanceof CertificateError) {
@@ -361,10 +363,10 @@ export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 	}
 
 	// See if the plugin client is authenticated.
-	const baseUrl = restClient.getAxiosInstance().defaults.baseURL;
+	const baseUrl = client.getAxiosInstance().defaults.baseURL;
 	if (baseUrl) {
 		storage.output.info(`Logged in to ${baseUrl}; checking credentials`);
-		restClient
+		client
 			.getAuthenticatedUser()
 			.then(async (user) => {
 				if (user && user.roles) {