+++ src/headers.ts
@@ -2,7 +2,7 @@ import * as cp from "child_process";
 import * as os from "os";
 import * as util from "util";
 import type { WorkspaceConfiguration } from "vscode";
-import { Logger } from "./logger";
+import { Logger } from "./logging/logger";
 import { escapeCommandArg } from "./util";
 
 interface ExecException {