+++ src/inbox.ts
@@ -1,14 +1,11 @@
-import { Api } from "coder/site/src/api/api";
 import {
 	Workspace,
 	GetInboxNotificationResponse,
 } from "coder/site/src/api/typesGenerated";
-import { ProxyAgent } from "proxy-agent";
 import * as vscode from "vscode";
-import { WebSocket } from "ws";
-import { coderSessionTokenHeader } from "./api";
-import { errToStr } from "./api-helper";
+import { CoderApi } from "./api/coderApi";
 import { type Storage } from "./storage";
+import { OneWayWebSocket } from "./websocket/oneWayWebSocket";
 
 // These are the template IDs of our notifications.
 // Maybe in the future we should avoid hardcoding
@@ -19,67 +16,39 @@ const TEMPLATE_WORKSPACE_OUT_OF_DISK = "f047f6a3-5713-40f7-85aa-0394cce9fa3a";
 export class Inbox implements vscode.Disposable {
 	readonly #storage: Storage;
 	#disposed = false;
-	#socket: WebSocket;
+	#socket: OneWayWebSocket<GetInboxNotificationResponse>;
 
-	constructor(
-		workspace: Workspace,
-		httpAgent: ProxyAgent,
-		restClient: Api,
-		storage: Storage,
-	) {
+	constructor(workspace: Workspace, client: CoderApi, storage: Storage) {
 		this.#storage = storage;
-
-		const baseUrlRaw = restClient.getAxiosInstance().defaults.baseURL;
-		if (!baseUrlRaw) {
-			throw new Error("No base URL set on REST client");
-		}
 
 		const watchTemplates = [
 			TEMPLATE_WORKSPACE_OUT_OF_DISK,
 			TEMPLATE_WORKSPACE_OUT_OF_MEMORY,
 		];
-		const watchTemplatesParam = encodeURIComponent(watchTemplates.join(","));
 
 		const watchTargets = [workspace.id];
-		const watchTargetsParam = encodeURIComponent(watchTargets.join(","));
 
-		// We shouldn't need to worry about this throwing. Whilst `baseURL` could
-		// be an invalid URL, that would've caused issues before we got to here.
-		const baseUrl = new URL(baseUrlRaw);
-		const socketProto = baseUrl.protocol === "https:" ? "wss:" : "ws:";
-		const socketUrl = `${socketProto}//${baseUrl.host}/api/v2/notifications/inbox/watch?format=plaintext&templates=${watchTemplatesParam}&targets=${watchTargetsParam}`;
+		this.#socket = client.watchInboxNotifications(watchTemplates, watchTargets);
 
-		const token = restClient.getAxiosInstance().defaults.headers.common[
-			coderSessionTokenHeader
-		] as string | undefined;
-		this.#socket = new WebSocket(new URL(socketUrl), {
-			agent: httpAgent,
-			followRedirects: true,
-			headers: token
-				? {
-						[coderSessionTokenHeader]: token,
-					}
-				: undefined,
-		});
-
-		this.#socket.on("open", () => {
+		this.#socket.addEventListener("open", () => {
 			this.#storage.output.info("Listening to Coder Inbox");
 		});
 
-		this.#socket.on("error", (error) => {
-			this.notifyError(error);
+		this.#socket.addEventListener("error", () => {
+			// Errors are already logged internally
 			this.dispose();
 		});
 
-		this.#socket.on("message", (data) => {
-			try {
-				const inboxMessage = JSON.parse(
-					data.toString(),
-				) as GetInboxNotificationResponse;
-
-				vscode.window.showInformationMessage(inboxMessage.notification.title);
-			} catch (error) {
-				this.notifyError(error);
+		this.#socket.addEventListener("message", (data) => {
+			if (data.parseError) {
+				this.#storage.output.error(
+					"Failed to parse inbox message",
+					data.parseError,
+				);
+			} else {
+				vscode.window.showInformationMessage(
+					data.parsedMessage.notification.title,
+				);
 			}
 		});
 	}
@@ -90,13 +59,5 @@ export class Inbox implements vscode.Disposable {
 			this.#socket.close();
 			this.#disposed = true;
 		}
-	}
-
-	private notifyError(error: unknown) {
-		const message = errToStr(
-			error,
-			"Got empty error while monitoring Coder Inbox",
-		);
-		this.#storage.output.error(message);
 	}
 }